requirements.txt
asgiref==3.7.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
Pygments==2.15.1
pyhcl==0.4.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pilmoji==2.0.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tzlocal==5.0.1
urllib3==1.26.16
webcolors==1.13
Werkzeug==2.3.4
zope.interface==6.0
